apps/backend/requirements.txt
anyio==3.6.1;   
cffi==1.15.1; 
click==8.1.3; 
colorama==0.4.5;
cryptography==37.0.4; 
dnspython==2.2.1;
ecdsa==0.18.0; 
email-validator==1.2.1; 
et-xmlfile==1.1.0; 
fastapi==0.79.0; 
h11==0.13.0; 
idna==3.3;   
openpyxl==3.0.10; 
passlib==1.7.4;
pyasn1==0.4.8;   
pycparser==2.21;    
pydantic==1.9.1; 
pymongo==4.2.0; 
python-dotenv==0.20.0; 
python-jose==3.3.0;
python-multipart==0.0.5;
rsa==4.9;   
six==1.16.0;    
sniffio==1.2.0;   
starlette==0.19.1;   
typing-extensions==4.3.0;   
uvicorn==0.18.2; 
